Time: N/A (0%), Space: N/A (0%) - LeetHub

+++ 2469-convert-the-temperature/2469-convert-the-temperature.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     vector<double> convertTemperature(double celsius) {
-        vector<double> ans;
+        vector<float> ans;
         ans.push_back(celsius + 273.15);
         ans.push_back(celsius * 1.80 +32.00);
         return ans;
